Fixed mask efficiency. Random number scales will be migrated to 0-1 rather than 0-100 to avoid these issues in the future. Mask efficiency didn't do anything until the model scaled up.  - It was set to a small number around 0.2 to 0.25 (looks like beta distributions with values 24.3 8.08)  - It was compared to random 100, ie there was only ever a 0.2% or so chance of a mask doing anything.  - Once the model scaled this error was fixed. Ie mask efficiency would now exist in a range of 20% to 25%.
Replaced infection transmission code. The old code was weird and copied to two places for some reason. It would sample a single simulant with the required properties sometimes, but in other times the random check was inside the simulant filter. The R and GlobalR update was in a completely separate part so I don't see how it could have been correlated to anything.